refactor(website): split FilesVisual into a folder structure
FilesVisual.tsx shell + barrel, with FileRow in components/, FileCategory
and FileEntry in types/, and FILE_ICONS / FILES in data/. RecordTabHeader
stays shared at the product-feature level.
