+++ packages/twenty-website/src/sections/product-feature/FilesVisual/data/files.ts
@@ -1,0 +1,9 @@
+import { type FileEntry } from '../types/file-entry';
+
+export const FILES: FileEntry[] = [
+  { category: 'document', date: 'Jul 1', name: 'NDA - Anthropic.pdf' },
+  { category: 'spreadsheet', date: 'Jul 12', name: 'Pricing Q4.xlsx' },
+  { category: 'presentation', date: 'Jul 18', name: 'Onboarding deck.pptx' },
+  { category: 'image', date: 'Jul 20', name: 'Brand assets.png' },
+  { category: 'document', date: 'Jul 24', name: 'Security review.docx' },
+];
