+++ packages/twenty-website/src/sections/product-feature/FilesVisual/index.ts
@@ -1,0 +1,1 @@
+export { FilesVisual } from './FilesVisual';
