+++ packages/twenty-website/src/sections/product-feature/FilesVisual/types/file-category.ts
@@ -1,0 +1,5 @@
+export type FileCategory =
+  | 'document'
+  | 'image'
+  | 'presentation'
+  | 'spreadsheet';
