+++ packages/twenty-website/src/sections/product-feature/FilesVisual/types/file-entry.ts
@@ -1,0 +1,7 @@
+import { type FileCategory } from './file-category';
+
+export type FileEntry = {
+  category: FileCategory;
+  date: string;
+  name: string;
+};
